auth/requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.109.0
orjson==3.9.12
pydantic==1.10.11
uvicorn==0.25.0
gunicorn==21.2.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.27
werkzeug==3.0.1
pyjwt==2.8.0
pytest==8.0.2
pytest_asyncio==0.23.5
alembic==1.13.1
email-validator==2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jaeger-client==4.4.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
httpx
yandexid
yandex-oauth
async-fastapi-jwt-auth
#pydantic-settings==2.0b2